+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Changed
 
+- React Native version to `0.73.4`
 - Update Bitmovin's native Android SDK version to `3.61.0`
 
 ### Fixed